build => delete image action

+++ actions/project.ts
@@ -6,7 +6,7 @@ import { isEqual } from "lodash"
 import { SchemaAddProject, TypeAddProject, TypeErrors, TypeReturnSererAction } from "./definition"
 import { ProjectCreateInput } from "@/lib/types"
 import { revalidatePath } from "next/cache"
-import { createImage, updateImage } from "./image"
+import { createImage, deleteImage, updateImage } from "./image"
 
 const projectObject = (formData: FormData) =>
   ({
@@ -17,7 +17,10 @@ const projectObject = (formData: FormData) =>
     description: formData.get("description"),
   } as TypeAddProject)
 
-export const addProject = async (formData: FormData): Promise<TypeReturnSererAction> => {
+export const addProject = async (
+  formData: FormData,
+  path: string
+): Promise<TypeReturnSererAction> => {
   const mainProject = projectObject(formData)
   const validationResult = SchemaAddProject.safeParse(mainProject)
 
@@ -35,6 +38,7 @@ export const addProject = async (formData: FormData): Promise<TypeReturnSererAct
     data: { ...validationResult.data, id: projectId, image: createImageResult.path },
   })
   if (projectResult) {
+    revalidatePath(path)
     return { message: "project create, successfully", status: true }
   }
 
@@ -81,12 +85,17 @@ export const editProject = async (
 }
 
 export const deleteProject = async (id: string, path: string): Promise<TypeReturnSererAction> => {
-  const deleteResult = await prisma.project.delete({ where: { id } })
-  if (deleteResult) {
-    revalidatePath(path)
-    return { message: "project removed successfully", status: true }
+  const projectResult = await prisma.project.findUnique({ where: { id } })
+  if (projectResult) {
+    const deleteImageResult = await deleteImage(projectResult.image)
+    const deleteResult = await prisma.project.delete({ where: { id } })
+    if (deleteResult && deleteImageResult.status) {
+      revalidatePath(path)
+      return { message: "project removed successfully", status: true }
+    }
+    return { message: "project remove got failure", status: false }
   }
-  return { message: "project remove got failure", status: false }
+  return { message: "project not found", status: false }
 }
 
 export const getAllProjects = async () => {
